Move isErrorLike to own file
