+++ src/errors/index.ts
@@ -1,7 +1,7 @@
-import type { ErrorLike } from "@apollo/client";
 import type { FetchResult } from "@apollo/client/link/core";
 
 import { CombinedProtocolErrors } from "./CombinedProtocolErrors.js";
+import { isErrorLike } from "./isErrorLike.js";
 import { UnconventionalError } from "./UnconventionalError.js";
 
 // This Symbol allows us to pass transport-specific errors from the link chain
@@ -24,17 +24,6 @@ export function graphQLResultHasProtocolErrors<T>(
     );
   }
   return false;
-}
-
-export function isErrorLike(error: unknown): error is ErrorLike {
-  return (
-    error !== null &&
-    typeof error === "object" &&
-    typeof (error as ErrorLike).message === "string" &&
-    typeof (error as ErrorLike).name === "string" &&
-    (typeof (error as ErrorLike).stack === "string" ||
-      typeof (error as ErrorLike).stack === "undefined")
-  );
 }
 
 export function toErrorLike(error: unknown) {
